Update Find to allow *
Has issues if the only argument is *, as the parser will evaluate that to every file in the current directory. Works with wrapping in '', but the user isnt expected to type that. Same issue with the directory if there is a space. Maybe update jsh to wrap the args from the command line.s